Add slot/props handling in codegen
Closes #466
