+++ .changeset/two-mice-grab.md
@@ -1,0 +1,5 @@
+---
+'@compai/css-gui': patch
+---
+
+Add prop syntax to relevant exports
